requirements.txt
azure-mgmt-network
azure-mgmt-resource
azure-mgmt-storage
azure-identity
